chore: rename selector-utils to utils
