+++ src/state/apis/swapper/selectors.ts
@@ -2,7 +2,7 @@ import { QueryStatus } from '@reduxjs/toolkit/dist/query'
 import type { SwapperName } from '@shapeshiftoss/swapper'
 import type { PartialRecord } from '@shapeshiftoss/types'
 import type { ReduxState } from 'state/reducer'
-import { createDeepEqualOutputSelector } from 'state/selectors/selector-utils'
+import { createDeepEqualOutputSelector } from 'state/selectors/utils'
 
 import type { TradeQuoteRequest } from './types'
