+++ src/state/selectors/search-selectors.ts
@@ -1,9 +1,6 @@
 import type { AssetId } from '@shapeshiftoss/caip'
 import type { Asset } from '@shapeshiftoss/types'
-import {
-  createDeepEqualOutputSelector,
-  selectSearchQueryFromFilter,
-} from 'state/selectors/selector-utils'
+import { createDeepEqualOutputSelector, selectSearchQueryFromFilter } from 'state/selectors/utils'
 import { selectAssetsBySearchQuery } from 'state/slices/assetsSlice/selectors'
 import { selectTxsByQuery } from 'state/slices/txHistorySlice/selectors'
 import type { TxId } from 'state/slices/txHistorySlice/txHistorySlice'
